#49 Refactor: 외부 utils.py 폐기 및 send_email 함수 이전
